requirements.txt
attrs==22.2.0
boto3==1.26.47
botocore==1.29.47
cached-property==1.5.2
cachetools==5.2.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
datapackage==1.15.2
db-dtypes==1.0.5
et-xmlfile==1.1.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.4.1
google-cloud-bigquery-storage==2.18.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
ijson==3.2.0.post0
isodate==0.6.1
jmespath==1.0.1
jsonlines==3.1.0
jsonpointer==2.3
jsonschema==4.17.3
linear-tsv==1.1.0
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-gbq==0.19.0
proto-plus==1.22.2
protobuf==4.21.12
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==2.0.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.46
tableschema==1.20.2
tabulator==1.53.5
unicodecsv==0.14.1
urllib3==1.26.14
xlrd==2.0.1
